fixed maybe camera reboot

+++ frost/frost/src/Libdc1394Grabber/Libdc1394Grabber.h
@@ -69,7 +69,7 @@ class Libdc1394Grabber : public ofxVideoGrabberSDK, public ofxThread
 		dc1394bayer_method_t bayerMethod;
 
 		dc1394camera_list_t*  list;
-	int blinkCounter;
+	unsigned long int blinkCounter;
 
 	private:
